Updated openstack/openstack Project: openstack/openstack-doc-tools  cba7f7f340876f6fc73867e8ee7751aaa6f7b7ef
Better identify deprecated configuration options

If a configuration option is tagged as deprecated in the code and the
help text does not start with DEPRECATED: add this to better identify
options that are no longer used.

If the code provides an additional deprecated reason, append this to
the provided help text of the configuration option.

Change-Id: I1a7c76c0109a6e59cc87d0ddc942c6d1cfb9dc91