Updated openstack/openstack Project: openstack/trove  a35c2e32572d185170c70d3f20ebdb438fbdeec3
Redis should perform backup using BGSAVE not SAVE

When Redis backup was implemented, using the save() command was
chosen as it blocks, making the code simpler to read and understand.
Unfortunately SAVE and BGSAVE have differences beyond one being
synchronous and the other asynchronous, in that SAVE also blocks
client connections while it runs. This behaviour is not optimal,
so the code was changed to use BGSAVE and wait until it completes.

In the case where bgsave() fails (possibly due to memory issues,
save() is called so that the backup will still complete.

The Redis helper was also modified to allow adding more data (to
increase the time the backup takes) and tests were rearranged
to improve chances of success.

Change-Id: I75d9df54b8550b08ce0112e516c17fd717ebb857
Closes-Bug: #1538334