Updated openstack/openstack Project: openstack/tempest  6322474e08ae4d13a29b4839c857c4b4f232185b
Fix API test for external subnet visibility

The test_external_network_visibility test incorrectly
assumes that subnets will never be visible on external
networks. However, this is not true in the case that a
network is both external and shared, in which case its
subnets will be shared and will be visible globally.

This updates the test to only make the subnet visibility
assertion on the configured public external network which
is stable and won't be in flux like networks created by
other tests.

At some point this test should just create its own external
network and make assertions on that rather than depend on
a specific configuration of the public network.

Related-Bug: #1553595
Change-Id: I0219e17d2ec70939039344826d35cb533b3ce065