Updated openstack/openstack Project: openstack/manila  9834c802b8bf565099abf357fe054e086978cf6e
HDS-HNAS: Fix exception in update_access not found

When share is deleted in the backend, user is unable to delete
share because of error thrown in update_access(). This
method throws a ShareResourceNotFound exception to be handled
by manager, thus allowing share to be deleted in this case,
without having to use force-delete.

Related-Bug: #1550377
Change-Id: I069ab60fef914dd6055f054dbbb777dc34b43b6f