Updated openstack/openstack Project: openstack-infra/project-config  47d7bd3afc500dae59e32516e983e16345e01bd4
Use pip installed dib for functional tests

With the switch to the ubuntu-trusty nodes, we've exposed a flaw in
testing dib from a checkout like this.  Several of the parts have
python dependencies [1] ; e.g.

 | dib-run-parts Thu Mar 17 06:04:57 UTC 2016
 |  Running /tmp/image.2NkOFQBR/hooks/extra-data.d/99-squash-package-install
 | Traceback (most recent call last):
 |   File "... /bin/package-installs-squash", line 23, in <module>
 |     import yaml
 | ImportError: No module named yaml

but we're not installing dib's python dependencies at all but simply
running ./run_tests.sh

So now that the image doesn't come with some of these things, we
break.  The solution is to install dib into a virtualenv and run from
there so that we get all the dependencies.  We already had an
experimental job doing this, so move it to the main job.  The scripts
should be using "#!/usr/bin/env python" so they should pickup the
virtualenv python & consequently have their dependencies.

[1] http://logs.openstack.org/33/293833/1/check/gate-dib-dsvm-functests-ubuntu-trusty/1731239/console.html

Change-Id: I5950c772b380cab926fd707cfed25db219a96f62