Updated openstack/openstack Project: openstack/nova  3ad46e3c7888f533745d4ba71fd3d4095d24e242
xenapi: Fix the live-migrate aggregate check

AggregateList now returns [] if the host is not in an
aggregate.  This was causing an IndexError in XenAPI.

This change also replaces the AggregateHostNotFound exception,
which is not caught properly further up the stack, with the
MigrationPreCheckError as this is a suitable substitute since
the aggregate check is only part of the pre migration check.

Probably introduced by Ibcb896b2eef673bce23491161ff394bcc87aa541

Related-Bug: #1557585
Change-Id: Icacd3cbe065ee8598e5e5f03bd73463b95ed2f75