Updated openstack/openstack Project: openstack/horizon  c45166c3fac46f2f4921a44c77e48f5b8f6847c5
Make ngcontainers the default Swift UI

Replace the old Swift UI with the new AngularJS implementation
by default.

Use setting in the enabled file to switch routing between the
two, allowing HEAT interface to seamlessly work with both.

Updated the release note with the new configuration information.

Removed the unnecessary additional "ngcontainers" panel.

Hard-coded the test suite to use the legacy panel - since they
are all testing the legacy panel and those tests will be removed
(or altered in the case of the stacks MappingTests) when
the legacy panel is removed.

Finally, I had to add a copyright notice to the enabled file
because flake8 noticed because I touched it or something. Very
random.

Change-Id: Ie4325185f57654522f9331d617f142687effb70b
Partially-Implements: blueprint angularize-swift
Co-Author: Travis Tripp <travis.tripp@hpe.com>
Co-Authored-By: Matt Borland <matt.borland@hpe.com>