Updated openstack/openstack Project: openstack/trove  5f0d88c163a3117ec259a33230ede003b1c3e1e2
Fix incorrect innobackupex args (fixed gate)

A recent release of Percona XtraBackup has tightened up the
validation of options passed to the innobackupex script.

See:
https://www.percona.com/doc/percona-xtrabackup/2.3/release-notes/2.3/2.3.4.html
https://bugs.launchpad.net/percona-xtrabackup/+bug/1533542

The options supplied to innobackupex in the MySQL restore strategy
are slightly incorrect - this was ignored previously but now causes
the backups to fail.

The options are now specified correctly.

Change-Id: Ic4671998f0c4467f17caa60cbb3106f49c505a2e
Closes-Bug: #1558794