Updated openstack/openstack Project: openstack/nova  2e2c225e831d68e75b0059ce0ec6383df3840868
Make Flavor.get operations prefer the API database

This makes the Flavor object get operations attempt to first load the
flavor from the api database, falling back to the main one if the
requested flavor is not found.

Note that I had to add backrefs for projects and extra_specs to the
Flavors model, which didn't have them yet. This is tested by the
code that tests loading from the actual DB models.

Related to blueprint flavor-cell-api

Co-Authored-By: Lalit Dagre <lalitdagre@gmail.com>
Change-Id: Ibc239a3d0ac0d22494919554024d1ba07a632218