Updated openstack/openstack Project: openstack/horizon  2187d2b0c31d4e7f34cb9ab50242fa0de41a14c9
Add Network Port selection to new instance launch

Adds a new step to the new instance launch
with the available ports.
The operator can add or remove ports to be
attached to the instance on boot.
This allows the operator to use ports
at launch instead or together with a network.
Adds tests and modifies other tests to take
into account the new number of steps and the
new addition to the model.

Change-Id: I29fdce433a5a8a9e3ebedbd1a1d96399890e4c6c
Implements: blueprint allow-launching-ports