Updated openstack/openstack Project: openstack/cinder  15528d110660216fd751fbb8dfe97cae3eef333c
Add get_volume_stats for LoggingVolumeDriver

The LoggingVolumeDriver will be useful in functional tests where
the backend volume being created is not important, for example the
nested quota Tempest tests. To allow volume creation, this driver
should report stats so that the scheduler will consider this backend
during volume creation.

Change-Id: I02fc02b7e4d9d6b7fec0a288fc34e2365ddcd9fd
Closes-Bug: #1554209