Updated openstack/openstack Project: openstack-infra/project-config  3a8fa2e3a2d04e771e12e7c891fbcbf7a3ddd7eb
Kolla docs-only patch -- take 2

This time with less gate breakage!

The goal, again, is to only run the docs gate when only docs have
changed. The regex has been validated against python 2.7.3 which
zuul runs. The last patch was validated on a newer python that
accepted the regex, but was run on an older version which broke zuul

Change-Id: Ic9214b1f22689940f62b6848728320f041dd6ad5
Related-Id: I0308428388b1f9621b7a7dd3aff214ae7b4460d9