Updated openstack/openstack Project: openstack/ironic  8e81b964a5fc815a7d42be749331571f4bb56be1
Extend the Conductor RPC object

This patch is extending the conductor RPC object adding a method for
registering and unregistering conductors. The touch() method was also
update to make the context parameter optional (since it's not actually
needed).

The base_manager.py module have been updated to use the new methods.

Change-Id: Ifb74909bc2cd39ce4ad056f7174285497e35035c