Updated openstack/openstack Project: openstack/nova  e11b6c06827a19e3079c798a4da9cdedd5b59e7d
Make Flavor create() and destroy() work against API DB

This makes Flavor.create() always create new flavors in the API DB,
and destroy delete from either (favoring the API DB first, like the
rest of the methods).

Now that we're this far, it also adds functional tests for the basic
operations of create, save, query, and destroy for flavors. It does
these both for flavors in the main database and the API database,
to prove that they work for both (except for create, which is always
against the API DB now).

Note that this also removes a couple of tests for reading deleted
flavors, which are now no longer legit cases across the board.

Related to blueprint flavor-cell-api

Change-Id: Id5fc2f95179beb9fd057a37744815a8e640e0c32