Updated openstack/openstack Project: openstack/nova-specs  a522e4b2d367d06103ca301d61005d9c6f26beb0
Detach and attach boot volumes

This spec was previously approved for Mitaka but was
delayed by volume-ops-when-shelved. That blueprint is
now complete.

It is sometimes useful for a cloud user to be able to
detach and attach the boot volume of an instance that
is not running. Currently nova does not allow this and
some operations assume it does not happen. This spec
proposes allowing the detach and attach of boot volumes
in the shelved_offloaded state and adding safeguards to
ensure it is safe.

blueprint detach-boot-volume

Previously-approved: Mitaka

APIImpact

Change-Id: I12e3953e1a238ee7e59d98a32b1b1c20b37ccd4c