Updated openstack/openstack Project: openstack/nova  7baf124ee77f958495cc8deedce9992d90bbfabc
Remove auto generated module api documentation

This disables the generation of the raw module api documentation in
our docs target. It is mostly not useful, as it builds a giant tree of
module documentation that are 98% boiler plate lists of methods with
no real content.

We leave the sphinx.ext.autodoc in the conf, which allows you to
specifically pull in modules for documentation when you want
to. doc/source/services.rst is an example of doing this in tree, which
still works after this change.

Change-Id: I4c10a8e45756cdcf612faca574e2fb3b7ffa2bdb