Updated openstack/openstack Project: openstack/cinder  1a31fd55d44834a64b7da90508e2593af0277b77
Fix error message when running genconfig

The Fujitsu eternus_dx_common code was configured to
try to import pywbem and immediately throw an error in
the case that the module wasn't available.  This would
cause the error to be reported during the genconfig process.

This patch changes the driver to just attempt to import
the module and record whether it was successful.  If not,
the error message isn't reported until __init__ is called.

This approach is consistent with what EMC does in their
drivers for importing pywbem.

Change-Id: I4f7d90bc2d05bde653f9d64fdb183ebf639fe3d4
Closes-bug: 1558644