Updated openstack/openstack Project: openstack/python-manilaclient  8fc8bd9eb24c664c87d57b6d5aefc6f1caa043f0
Revert fix access rules functional tests

When update_access was first implemented in manila, it merged with
a bug that was not allowing rules to be added when
access_rules_status were 'out_of_sync'. Due to this limitation,
manila client functional tests were fixed. Now, since the initial
bug was fixed on manila, this can be reverted.

Change-Id: I4aee2d339ec96e44208771c19714f9911115a81a
Depends-On: I0f863cbae4d8af0660114161deda7bf7aa60d71d
Closes-Bug: #1556956