Updated openstack/openstack Project: openstack-infra/shade  e1f6f1c9eea2d6b0bf5b32a7c16630e233ca6cdb
Run extra specs through TaskManager and use requests

The previous fix for flavors and extra_specs made API calls that did not
get wrapped in the TaskManager. This is really novaclient's fault for
having a weird interface. Change it to call in through the api.client
interface on the flavor manager in novaclient so we can Task it.

Some clouds return the extra_specs dict in a key called
OS-FLV-WITH-EXT-SPECS:extra_specs. If that's present, just add it to the
return dict instead of making the extra API call.

There are two instances of making raw REST calls to nova via the
client.api property of novaclient. However, we have access to a proper
keystoneauth Session object we can mount properly on the compute
service, so instead of using weird internal novaclient API - let's just
use requests.

In order to support this, we need to do the request_id logic ourselves.
Luckily for us it's easier logic than in python-*client because we don't
have to deal with appending request ids since we're never in the middle
of a request chain.

Change-Id: Ia8e40d542f98f0ccc55131ca277b862cffeecb12