Updated openstack/openstack Project: openstack/horizon  0ab96ce8f44720734e96875d0eb1188845259a0f
Fixing test cases

The tests in ceilometer_tests.py for fetching the meter list,
alarm list, sample list, statistic list and the resource list
are incomplete.It verifies the API call, but the API could fail
completely and an empty list would still pass the assertion.
This patch fixes it.

Change-Id: I532992a9a7ad150c83e7f18721d885721070b029
Closes-bug:#1536470