Updated openstack/openstack Project: openstack/heat  ff75d3a111e1904d44fa9d1a5549c08eda5d605a
Remove unknown parameters in patched update

When updating a stack with the existing flag, we keep the parameters
from the old template to be used against the new version. Sometimes
parameters will get remove and won't make sense anymore, and keeping
them would break update with a 'Parameter was not defined' error. This
filters out such parameters so that the updates succeed.

Change-Id: I6f2aa77da28d271dd001a137bb574b5470292f15
Closes-Bug: #1558610