Updated openstack/openstack Project: openstack/swift  919608caaf1852661aa2809b75817a845c0130ee
Don't report recon mount/usage status on files

Today recon will include normal files in the payload it returns for
/recon/unmounted and /recon/diskusage. As a result it can trigger
bogus alarms on any operations-side monitoring checking for unmounted
disks or disks that show up in diskusage with weird looking stats.

This change adds an isdir check for the entries it finds in /srv/node.

Change-Id: Iad72e03fdda11ff600b81b4c5d58020cc4b9048e
Closes-bug: #1556747