Updated openstack/openstack Project: openstack/tripleo-puppet-elements  54ab4846892631d0d64d5eafbc1fa6529c14a34f
Add rubygems-deep_merge to package list

The deep merge gem enables hiera to be able to merge hashes more
thoroughly than just the top level hashes. This is useful if we want
to add extra configurations in hiera that are already being
configured elsewhere in hiera (such as
keystone::config::keystone_config). Because currently what will
happen, for instance, with the keystone_config, is that it will take
the value of the hash that is defined later in the hierarchy.

Change-Id: Ib855a287491bf36e2609b354f804923f83ca2a1c