Updated openstack/openstack Project: openstack/python-heatclient  f13d36ae0114e159663ea2a3fa07260df9d9407d
Fix regression handling relative paths in environment

7627e1bae53747e1471caf632dfc1b6edf2e26cf introduced new code to add the
environment to the files mapping, so server side resolution of multiple
environments is possible.

However, the environment added is wrong when you're handling those with
relative paths inside, we must add the resolved version or the keys
don't map to keys in the files map, an you get an error when we try
to locate the file on the server.

Evidently we lack adequate functional test coverage of this, which can
be addressed via a heat patch, but I added a unit test illustrating
the problem.

Change-Id: I3bd2cd48623e4ad1b4067f4cefc7ddff30b9e88d
Closes-Bug: #1560106