Updated openstack/openstack Project: openstack/cinder  545b8c4435469c72920d54e4a3ca3c74cbc66b95
Report 2.0 as supported version of RPC APIs

When bumping major versions of RPC APIs from 1.x to 2.0 I haven't raised
the version the managers are reporting into the `services` table.
Though this doesn't have any immediate consequences, it effectively
blocks us from dropping 1.x support in Newton, because we need that
support to downgrade the messages to what Mitaka reports.

This commit fixes that problem by making the Mitaka's services to report
2.0 as RPC API version.

Closes-Bug: 1560062

Change-Id: I45e695025f1cb38fd8b062593b29da7463ba22fc