Updated openstack/openstack Project: openstack/neutron  53debf1b22e8111a0d57bedf78bf29085c543d71
Enforce alembic branch when creating scripts manually

Autogeneration code already enforces the branch for new scripts, but
explicit mode [where you pass --expand or --contract to get empty script
stubs] was not doing it. This resulted in new scripts being created in
the same directory where the current branch head is located. This is a
problem when we want to switch to a new release directory to store
future scripts.

To fix the problem, we now enforce the branch as per --contract/--expand
CLI arguments passed. We also ensure that the new directory is present
before attempting to generate a new stub script.

Related-bug: #1501380
Change-Id: I01cb6fed6dd246bedf945c323372e53991973fdc