Updated openstack/openstack Project: openstack/nova  3b2895b960630934f827918d50577f9167110cab
VMware: use datacenter path to fetch image

When fetching image on nova using VMware VC driver,
datacenter name is passed instead of using complete
path to the datacenter. The incorrect url generated
causes error in downloading the image as reported in
Bug #1369019. This patch uses the complete
inventory path of the datacenter to fetch the image.

Change-Id: I88bfe37bffb4dc38eb27472167753d1c28db7a97
Closes-Bug: #1369019