Updated openstack/openstack Project: openstack/neutron  0a92395f6524ecb82637ddef9ff75f66cc755d69
Fix wrong use of list of dict in _check_driver_to_bind

From [1], the segments_to_bind should be a list of dict, so the
"level.segment_id in segments_to_bind" will never work.

This patch extracts a set of segment ids and uses the set in the
if condition.

[1] https://goo.gl/yKYSTA

Change-Id: I58f1d128e6cd79546d84f7d5bfcb026affc4fc5e
Closes-bug: #1524356