Updated openstack/openstack Project: openstack-infra/project-config  2df8a6a0720647e5e542d8c35c56c4d834132a47
keystone: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I0d7a5a3d6cc667e2e849fa8a934e70ad0e61c6e9
Needed-By: I5b691d0adc812d687b318f639544ecaa165b9cfc