Updated openstack/openstack Project: openstack-infra/project-config  bb0b9e69dff9a30d7c4f9f4513ffaec3d138d639
ironic,ironic-inspector: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I6056603fed09a486789808e3097d4c2a44a1acbe