Updated openstack/openstack Project: openstack/neutron  3d6cb95f1932638e4c0772979d6e20b34de5f601
Removes host file contents from DHCP agent logs

The change I3ad7864eeb2f959549ed356a1e34fa18804395cc addressed a
performance impact by removing logging calls from inside the loop that
creates the hosts file in memory. However, it also introduced the
dumping of said hosts file to the logs.

On deployments with a very high number of ports(>5K), this increases the
load of the node substantially. The problem is worse when a high number
of port updates is received in a short amount of time.

There is also the administrative burden caused by the sheer amount of
data being logged.

If the content of the host file is required for debugging purposes, it
can be found at /opt/stack/data/neutron/dhcp/<net-id>/host

This patch is trivial and only removes the logging of the file contents.
The 'building'/'done building' logging calls may still provide useful
information.

Change-Id: Ie176ef123c194c22dca0967a6acfb9a48c959e6c
Closes-Bug: 1557640