Updated openstack/openstack Project: openstack-infra/project-config  7bc89f87827c07c32878a1a182e430a8e2db98c3
oslo.db: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I8dac66bda4f28606f7009a273425b6d7f1869910