Updated openstack/openstack Project: openstack-infra/project-config  cc26bb2b0c1677affa8fd078c7fad1677e03d3f0
storyboard: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I413a2e181f3ff2741b03f6a658af8af5bb8a11b0