Updated openstack/openstack Project: openstack-infra/project-config  11a8ceec8a4bb993ceda9df1c53d7ac927205e6d
taskflow: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: I5b8c0292db39bc16de83e49acca6771d0bebd3d1