Updated openstack/openstack Project: openstack-infra/project-config  c10ab59ba4f8e212fbe149bd0e0bddeea450220e
nodepool: Use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
This repo needs databases setup for unit tests.

Change-Id: Id42e5bc868b424e0a07ce09a7755c2acf1f2f949