Updated openstack/openstack Project: openstack/swift  d08ed1011d314e4d99abb21ec4e7d622b2ac827a
Set backend content length for fallocate - EC Policy

Currently, the ECObjectController removes the 'content-length' header.
This part is ok, except that value is being used to set
'X-Backend-Obj-Content-Length', so it is always 0.  This leads to not
calling fallocate (details on bug) on a PUT since the size is 0.

This change makes use of some numbers returned from the EC Driver
get_segment_info method in order to calculate the expected on-disk
size that should be allocated.  The EC controller will now set the
'X-Backend-Obj-Content-Length' value appropriately.

Co-Authored-By: Kota Tsuyuzaki
Co-Authored-By: John Dickinson
Co-Authored-By: Tim Burke
Change-Id: Ifd16c1438539e6fd9bb2dbcd053d11bea2e09fee
Fixes: bug 1532008