Updated openstack/openstack Project: openstack-dev/grenade  575b2cc9565cdd9af78a692dc1b5fc1d46fdef8a
Run online data migrations for nova

Note that this is actually only required when going from mitaka->newton,
but in order to land things very early in newton, we need this to be
run. Since it's valid in mitaka too, we can go ahead and land it there,
and then copy or move it to from-mitaka later.

Change-Id: Ibc02acc60b1023039f2613f8949b95d55169fd30