Updated openstack/openstack Project: openstack-infra/project-config  da14a3c716fd38b5672bfa37583a9e1cb8126f47
net-bgpvpn: control gate settings from rc scripts

This change updates jenkins job definitions for
networking-bgpvpn so that gate settings can be controlled
from rcfiles in the networking-bgpvpn repo, instead
of being hardcoded.

This change removes a few lines because they now
are present in the rcfile that this change enables.

Depends-On: I8d9580b480b1cdd3d185d51f13cd06be26310d17
Change-Id: Ie99cb55febda08c2f16604cc90455764a7aa3a80